pyproject.toml
[project]
name = "bianlunmiao-backend"
version = "0.1.0"
description = "辩论喵后端 MVP 1.0 (FastAPI)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.14.0",
  "psycopg2-binary>=2.9.10",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.6.1",
  "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "httpx>=0.28.1",
  "ruff>=0.8.4",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]
